[test-suite] Fix cycle imports
